+++ packages/checkbox/src/index.tsx
@@ -1,1 +1,2 @@
-export { Checkbox as default, CheckboxProps } from './Checkbox';
+export { Checkbox, CheckboxProps } from './Checkbox';
+export { default } from './Checkbox';