+++ packages/date/src/DateRange.tsx
@@ -1,7 +1,8 @@
 import { FC, HTMLAttributes } from 'react';
 import { DateRangeElement, DateRangeElementProps } from '@equinor/fusion-wc-date';
 import { elementAttributes } from '@equinor/fusion-react-utils';
-export { DateTimeFormat } from '@equinor/fusion-wc-date';
+
+DateRangeElement;
 
 export type DateRangeProps = DateRangeElementProps &
   HTMLAttributes<DateRangeElement> & {
@@ -10,9 +11,6 @@ export type DateRangeProps = DateRangeElementProps &
 
 export const DateRange: FC<DateRangeProps> = (props) => {
   const attributes = elementAttributes<DateRangeProps>(props);
-
-  console.log(attributes);
-
   return <fwc-daterange {...attributes}></fwc-daterange>;
 };