+++ packages/table/src/plugins/menu/reducer.ts
@@ -13,13 +13,10 @@ export const reducer = <D extends TableData>(
       const { menu } = state;
       const { columnId } = payload;
 
-      // console.log(state, payload);
-
       if (columnId === menu.columnId) {
         const show = payload.show === undefined ? !menu.show : payload.show;
         return { ...state, menu: { columnId, show } };
       } else if (payload.show !== false) {
-        console.log('ok');
         return { ...state, menu: { columnId, show: true } };
       }