+++ packages/checkbox/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.3.1 (2021-09-07)
+
+**Note:** Version bump only for package @equinor/fusion-react-checkbox
+
+
+
+
+
+# 0.3.0 (2021-09-07)
+
+
+### Bug Fixes
+
+* add label to checkbox ([a752f4f](https://github.com/equinor/fusion-react-components/commit/a752f4f64616d365f8046d9ee213c3373c3964c3))
+* fix exports and deps ([b614834](https://github.com/equinor/fusion-react-components/commit/b614834c32db4fbb9b06407e53557109128ec95b))
+* import of elements ([b19fc55](https://github.com/equinor/fusion-react-components/commit/b19fc55471dd4dc7d44e2216699d1d12979d4211))
+
+
+### Features
+
+* update doc and include readme ([04b2544](https://github.com/equinor/fusion-react-components/commit/04b25443398507b35c3b88bf90a26d56c5b1c460))
+
+
+
+
+
 ## 0.2.2 (2021-08-03)